// src/events/control/ControllerEvent.d.ts
import ReadStream from "../../streams/ReadStream";
import { StatusBytes } from "../../streams/StatusBytes";
import WriteStream from "../../streams/WriteStream";
import ControlEvent from "./ControlEvent";
export declare enum ControllerType {
    BANK_SELECT_COARSE = 0,
    MODULATION_WHEEL_COARSE = 1,
    BREATH_CONTROLLER_COARSE = 2,
    FOOT_CONTROLLER_COARSE = 4,
    PORTAMENTO_TIME_COARSE = 5,
    DATA_ENTRY_COARSE = 6,
    CHANNEL_VOLUME_COARSE = 7,
    BALANCE_COARSE = 8,
    PAN_COARSE = 10,
    EXPRESSION_COARSE2 = 11,
    EFFECT_CONTROL_1_COARSE = 12,
    EFFECT_CONTROL_2_COARSE = 13,
    GENERAL_PURPOSE_CONTROLLER_1_COARSE = 16,
    GENERAL_PURPOSE_CONTROLLER_2_COARSE = 17,
    GENERAL_PURPOSE_CONTROLLER_3_COARSE = 18,
    GENERAL_PURPOSE_CONTROLLER_4_COARSE = 19,
    BANK_SELECT_FINE = 32,
    MODULATION_WHEEL_FINE = 33,
    BREATH_CONTROLLER_FINE = 34,
    FOOT_CONTROLLER_FINE = 36,
    PORTAMENTO_TIME_FINE = 37,
    DATA_ENTRY_FINE = 38,
    CHANNEL_VOLUME_FINE = 39,
    BALANCE_FINE = 40,
    PAN_FINE = 42,
    EXPRESSION_FINE2 = 43,
    EFFECT_CONTROL_1_FINE = 44,
    EFFECT_CONTROL_2_FINE = 45,
    HOLD_PEDAL_1 = 64,
    PORTAMENTO_PEDAL = 65,
    SOSTENUTO_PEDAL = 66,
    SOFT_PEDAL = 67,
    LEGATO_PEDAL = 68,
    HOLD_PEDAL_2 = 69,
    SOUND_CONTROLLER_1 = 70,
    SOUND_CONTROLLER_2 = 71,
    SOUND_CONTROLLER_3 = 72,
    SOUND_CONTROLLER_4 = 73,
    SOUND_CONTROLLER_5 = 74,
    SOUND_CONTROLLER_6 = 75,
    SOUND_CONTROLLER_7 = 76,
    SOUND_CONTROLLER_8 = 77,
    SOUND_CONTROLLER_9 = 78,
    SOUND_CONTROLLER_10 = 79,
    GENERAL_PURPOSE_CONTROLLER_5 = 80,
    GENERAL_PURPOSE_CONTROLLER_6 = 81,
    GENERAL_PURPOSE_CONTROLLER_7 = 82,
    GENERAL_PURPOSE_CONTROLLER_8 = 83,
    PORTAMENTO_CONTROL = 84,
    HIGH_RESOLUTION_VELOCITY_PREFIX = 88,
    EFFECT_1_DEPTH = 91,
    EFFECT_2_DEPTH = 92,
    EFFECT_3_DEPTH = 93,
    EFFECT_4_DEPTH = 94,
    EFFECT_5_DEPTH = 95,
    DATA_BUTTON_INCREMENT = 96,
    DATA_BUTTON_DECREMENT = 97,
    NON_REGISTERED_PARAMETER_COARSE = 98,
    NON_REGISTERED_PARAMETER_FINE = 99,
    REGISTERED_PARAMETER_COARSE = 100,
    REGISTERED_PARAMETER_FINE = 101,
    ALL_SOUND_OFF = 120,
    ALL_CONTROLLERS_OFF = 121,
    LOCAL_CONTROL = 122,
    ALL_NOTES_OFF = 123,
    OMNI_MODE_OFF = 124,
    OMNI_MODE_ON = 125,
    MONO_OPERATION_AND_ALL_NOTES_OFF = 126,
    POLY_OPERATION_AND_ALL_NOTES_OFF = 127
}
export default class ControllerEvent extends ControlEvent {
    controller: ControllerType;
    value: number;
    readBytes(stream: ReadStream): void;
    writeBytes(stream: WriteStream, status?: StatusBytes): void;
    protected getTypeHibyte(): number;
}